First pass at new main memory model. The latency has changed by approximately 1 clock cycle, so adjustments to configuration parameters may be required to have comparable results. The bandwidth for multiple tiles is now reduced to a realistic figure, so multitile benchmarks may behave differently.

+++ src/TileComponents/Memory/MagicMemory.cpp
@@ -49,15 +49,15 @@ void MagicMemory::operate(MemoryOpcode opcode,
     }
 
     case STORE_W:
-      mainMemory.writeWord(address, data);
+      mainMemory.writeWord(address, data.toUInt());
       break;
 
     case STORE_HW:
-      mainMemory.writeHalfword(address, data);
+      mainMemory.writeHalfword(address, data.toUInt());
       break;
 
     case STORE_B:
-      mainMemory.writeByte(address, data);
+      mainMemory.writeByte(address, data.toUInt());
       break;
 
     case IPK_READ: {
